requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
python-dotenv
qrcode[pil]
pandas
openpyxl
gunicorn
psycopg2-binary 